Upgraded IDs validation of streams

+++ src/main/java/store/RedisStore.java
@@ -6,7 +6,7 @@ import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.ConcurrentLinkedQueue;
 
 public class RedisStore {
-    public  Integer milisecondsTime = 0;
+    public  Long milisecondsTime = (long) 0;
     public  Integer sequenceNumber = 0;
     public  final ConcurrentHashMap<String , RedisData> data = new ConcurrentHashMap<>();
     public  final ConcurrentHashMap<String , ArrayDeque<String>> list = new ConcurrentHashMap<>();
